change all code mentions of splits to feature flags
