+++ nextjs-feature-flags-middleware/lib/posthog.ts
@@ -1,12 +1,12 @@
 /**
- * List of known active Splits
+ * List of known active Feature Flags
  */
-export const SPLITS = {
+export const FEATURE_FLAGS = {
   NEW_ABOUT_PAGE: 'New_About_Page',
   NEW_MARKETING_PAGE: 'New_Marketing_Page',
 } as const
 
-export type SPLITS = typeof SPLITS[keyof typeof SPLITS]
+export type FEATURE_FLAGS = typeof FEATURE_FLAGS[keyof typeof FEATURE_FLAGS]
 
 /**
  * Sends an event to Split.io
