+++ nextjs-feature-flags-middleware/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "feature-flag-split",
+  "name": "posthog-feature-flags",
   "version": "1.0.0",
   "repository": "https://github.com/vercel/examples.git",
   "license": "MIT",
