+++ nextjs-feature-flags-middleware/pages/marketing/_middleware.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { isFeatureFlagEnabled } from '@lib/split-node'
-import { SPLITS } from '@lib/split'
+import { isFeatureFlagEnabled } from '@lib/posthog-node'
+import { FEATURE_FLAGS } from '@lib/posthog'
 
 export async function middleware(req: NextRequest) {
   // Redirect paths that go directly to the variant
@@ -8,10 +8,10 @@ export async function middleware(req: NextRequest) {
     return NextResponse.redirect('/marketing')
   }
 
-  const flagName = `flag-${SPLITS.NEW_MARKETING_PAGE}`
+  const flagName = `flag-${FEATURE_FLAGS.NEW_MARKETING_PAGE}`
   const cookie =
     req.cookies[flagName] ||
-    (await isFeatureFlagEnabled(Math.random().toString(), SPLITS.NEW_MARKETING_PAGE)? '1' : '0')
+    (await isFeatureFlagEnabled(Math.random().toString(), FEATURE_FLAGS.NEW_MARKETING_PAGE)? '1' : '0')
 
   const res = NextResponse.rewrite(
     cookie === '1' ? '/marketing/b' : '/marketing'
